feat: Command Pattern 구현 > Command 패키지를 Service 패키지로 이동
