+++ CommandPattern/src/service/WalkCommand.java
@@ -1,17 +1,17 @@
-package domain.command;
+package service;
 
 import domain.Player;
 
-public class FlyCommand implements Command {
+public class WalkCommand implements Command {
 
     private final Player player;
 
-    public FlyCommand(Player player) {
+    public WalkCommand(Player player) {
         this.player = player;
     }
 
     @Override
     public void execute() {
-        player.fly();
+        player.walk();
     }
 }
